+++ lynglib/src/commonMain/kotlin/net/sergeych/lyng/obj/ObjInstanceClass.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Sergey S. Chernov real.sergeych@gmail.com
+ * Copyright 2026 Sergey S. Chernov real.sergeych@gmail.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,13 +28,26 @@ import net.sergeych.lynon.LynonType
 class ObjInstanceClass(val name: String, vararg parents: ObjClass) : ObjClass(name, *parents) {
 
     override suspend fun deserialize(scope: Scope, decoder: LynonDecoder, lynonType: LynonType?): Obj {
-        val args = decoder.decodeAnyList(scope)
-        val actualSize = constructorMeta?.params?.size ?: 0
-        if (args.size > actualSize)
-            scope.raiseIllegalArgument("constructor $name has only $actualSize but serialized version has ${args.size}")
-        val newScope = scope.createChildScope(args = Arguments(args))
-        val instance = createInstance(newScope)
-        initializeInstance(instance, newScope.args, runConstructors = false)
+        val serializedArgs = decoder.decodeAnyList(scope)
+        val meta = constructorMeta ?: scope.raiseError("no constructor meta for $name")
+        val nonTransientCount = meta.params.count { !it.isTransient }
+        if (serializedArgs.size != nonTransientCount)
+            scope.raiseIllegalArgument("constructor $name expects $nonTransientCount non-transient arguments, but serialized version has ${serializedArgs.size}")
+        
+        var sIdx = 0
+        val namedArgs = mutableMapOf<String, Obj>()
+        for (p in meta.params) {
+            if (!p.isTransient) {
+                namedArgs[p.name] = serializedArgs[sIdx++]
+            } else if (p.defaultValue == null) {
+                // If transient parameter has no default value, we use ObjNull to avoid "too few arguments" error
+                namedArgs[p.name] = ObjNull
+            }
+        }
+        // Using named arguments allows the constructor to apply default values for transient parameters
+        val newArgs = Arguments(list = emptyList<Obj>(), named = namedArgs)
+        val instance = createInstance(scope.createChildScope(args = newArgs))
+        initializeInstance(instance, newArgs, runConstructors = false)
         return instance.apply {
             deserializeStateVars(scope, decoder)
             invokeInstanceMethod(scope, "onDeserialized") { ObjVoid }